# OBJ loader optimization