This PR solves #8 and introduces a more efficient OBJ loading, which is especially faster on smaller meshes.

# (Preliminary) Loading times

No actual statistics, just examples.

## MacBook Pro 2014

### After optimization

* Powerplant: 97854.898825 ms
* Rungholt: 50352.615261 ms
* Sponza: 2272.888937 ms

## Xeon Workstation

### Before optimization

* Powerplant: 55771ms
* Rungholt: 27734ms
* Sponza: 1593ms

### After optimization

* Powerplant: 67706ms
* Rungholt: 32548ms
* Sponza: 1200ms

## MeshLab

* Powerplant: 73000ms
* Rungholt: 33260ms
* Sponza: (Crashes)